Refactor UserHomepage to Use TopBar
